+++ src/main/java/com/csec/CatholicTableMatching/controller/MatchingController.java
@@ -1,10 +1,12 @@
 package com.csec.CatholicTableMatching.controller;
 
-import com.csec.CatholicTableMatching.domain.User;
 import com.csec.CatholicTableMatching.domain.MatchForm;
+import com.csec.CatholicTableMatching.security.domain.User;
+import com.csec.CatholicTableMatching.security.repository.UserRepository;
 import com.csec.CatholicTableMatching.service.MatchingService;
-import com.csec.CatholicTableMatching.service.UserService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
@@ -15,7 +17,7 @@ public class MatchingController {
 
     private MatchingService matchingService;
 
-    private UserService userService;
+    private UserRepository userRepository;
 
     private MatchForm matchForm;
 
@@ -25,20 +27,24 @@ public class MatchingController {
     }
 
     @GetMapping("/match")
-    public String MatchStart(@ModelAttribute User user ,Model model){
+    @PreAuthorize("isAuthenticated()")
+    public String MatchStart(@ModelAttribute User user , Model model){
         model.addAttribute("matchForm", new MatchForm());
         return "match_form_nes";
     } // 매치 폼 화면이동
 
+    @PreAuthorize("isAuthenticated()")
     @PostMapping("/match")
     public String findMatches(@ModelAttribute("matchForm") MatchForm matchForm) {
         Long userId = matchForm.getUser().getId(); // 가정: getUser()가 사용자 ID를 반환한다.
         return "redirect:/match/" + userId;
     } // 매치 넣으면 매치현황판 으로 이동할수 있게
 
+    @PreAuthorize("isAuthenticated()") // todo 사용자 id와 인증한 주체의 id가 동일한지도 검사 필요
     @RequestMapping("/match/{userId}")
     public String MatchStatus(@PathVariable("userId") Long userId, Model model){
-        User user = userService.findUserById(userId);
+        User user = userRepository.findUserById(userId).orElseThrow(
+                () -> new RuntimeException("No customer found with ID " + userId));
         model.addAttribute("customer", user);
         return "match_status";
