+++ src/main/java/com/csec/CatholicTableMatching/security/oauth/PrincipalDetails.java
@@ -32,7 +32,7 @@ public class PrincipalDetails implements UserDetails, OAuth2User {
 
     @Override
     public String getPassword() {
-        return user.getPassword();
+        return null;
     }
 
     @Override
