+++ src/main/resources/application.yml
@@ -9,7 +9,6 @@ spring:
     password:
     driver-class-name: org.h2.Driver
 
-
   h2:
     console:
       enabled: true
@@ -22,7 +21,6 @@ spring:
       hibernate:
         show_sql: true
         format_sql: true
-
 
 logging.level:
   org.hibernate.SQL: debug
