+++ src/main/resources/templates/index.html
@@ -61,7 +61,7 @@
         <section class="icon-list">
             <i class="nes-kirby"></i>
         </section>
-        <a th:href="@{/login}" class="nes-btn is-primary">로그인</a>
+        <a th:href="@{/oauth2/authorization/kakao}" class="nes-btn is-primary">카카오로 로그인</a>
     </div>
 </div>
 </body>
